score 'double'
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@5343 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/web/api/ApiContent.java
@@ -2544,7 +2544,7 @@ public class ApiContent extends ApiGeneric {
             episode = episodeMap.get(program.getEpisodeId());
             if (episode != null) {
                 String counterName = "s_ch" + episode.getChannelId() + "_e" + episode.getId();
-                double score = factory.getCount(counterName) / 1000000;
+                double score = (double)factory.getCount(counterName) / 1000000;
                 log.info("counter name = " + counterName);
                 log.info(episode.getName() + ", score = " + score);
                 result.put("score", score); // score: 得分
